Refactor and add descriptor read
Refactor to make BleXXXXX object creation more general with generic
types for library specific types in the hope that this might actually
support more than one underlying library eventually.

Add descriptor read which is restricted by a mutex to a single read at
a time since there seem to be problems with multiple reads of different
services even on the same device.

Not sure if that's a bug or a feature of the underlying package or
native implementation.