Use User Description descriptor for name
If a characteristic doesn't have a name try to get a User Description
descriptor value to use.

Refactored a lot of the bleXXXProviders to class format when trying
to fix stuff and now user listen instead of watch.

Not clear how much of that was necessary but it's working so I'll let
it be for now.