Add bleCharacteristicNotification along with some refactoring
Create bleCharacteristicNotification which returns the raw values
for a characteristic using notification.

Doesn't use a stream since I'm not sure what the benefit would be and
the code that gets the values is deep in the listen nesting.

Also some refactoring of to get characteristic along with some effort
at better error handling with my own exception class heirarchy.