Handle values in exceptionMessage to be a function of one parameter
If the value in a MsgXXXList is a function call it with the name as
the parameter and use the result as the object.

Format value now receives the value 'name' as the first parameter so
it can be passed to the function or used by override for any purpose
that seems useful.