Create quick_blue variant of Ble with scanner implemented
Create the quick_blue variant and implement the scanner code along with
an init method to setup the backend and the logging with some default
values.

Works on both Android and IOS but no bueno getting any scan results back
on MacOS which is the same result as flutter_blue_plus. Puzzling!