Major reworking of bleExceptionDisplayMessage
Try to return something more informative without the messiness of the
whole stack by combining the error message itself along with the
root of the causedBy chain so we display what we were trying to do
along with the deepest exception we can find which is as close to the
source of the problem that we can access.

Will this be sufficient and useful or is further work needed?