Always discoverServices after connecting for flutter_blue_plus
Version 1.28.9 no longer clears knownServices on connect (or disconnect)
so we'll execute discoverServices when we connect rather than trying to
keep track of whether this is a disconnect.

We'll need to do it anyway at some point (probably).

Also accommodate the breaking change for 1.28.10 never returning null
for servicesList. Hopefully they won't change this again.