Refactor characteristic handling and add web characteristic/descriptor
Characteristics were part of BleService which makes sense on those
platforms which fetched them with the service but that's not always the
case and it breaks the model of referring to things by the set of
qualifying properties.

Remove the array of BleCharacteristic from BleService and add additional
providers to explicitly fetch them when wanted.

Add web support for characteristic and descriptors using this new model.