Add initialization parameter exitWhenRequested
The exitWhenRequested parameter is only used by Linux backend
to give the library an opportunity to close all the open devices
when shutting down the application since it doesn't appear that
BlueZ will do it by itself.

If set to false it will be the responsibility of the application
to call riverpodBleDispose before exiting or existing connections
to devices may be left open.